+++ inbm/Changelog.md
@@ -8,12 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## Fixed
  - HSD 15010509095 - INBC fail to return correct Exit Code for difference scenario (-4/-6/-11)
  - HSD 15010982715 - inbm-vision failed to receive query request from cloud
+ - HSD 15011009937 - Remove POTA Failure error on Success
 
 ### Added
 - RTC 508495 Support INBC Config Append command
+- RTC 508497 Support INBC Config Remove command
 
 ## 3.0.10 - 2022-03-28
-NOTE: This release has not yet been security tested.
 
 ## NEXT - 3.0.9 - 2022-03-08
  - RTC 509991 Updated docker stats dashboard to display units for readability.